Handle delete temporary type

+++ packages/strapi-plugin-content-type-builder/admin/src/containers/DataManagerProvider/index.js
@@ -196,9 +196,24 @@ const DataManagerProvider = ({ allIcons, children }) => {
   const deleteData = async () => {
     try {
       const requestURL = `/${pluginId}/${endPoint}/${currentUid}`;
+      const isTemporary = get(
+        modifiedData,
+        [firstKeyToMainSchema, 'isTemporary'],
+        false
+      );
 
       // Close the modal
       push({ search: '' });
+
+      if (isTemporary) {
+        // Delete the not saved type
+        // Here we just need to reset the components to the initial ones and also the content types
+        // Doing so will trigging a url change since the type doesn't exist in either the contentTypes or the components
+        // so the modified and the initial data will also be reset in the useEffect...
+        dispatch({ type: 'DELETE_NOT_SAVED_TYPE' });
+
+        return;
+      }
 
       await request(requestURL, { method: 'DELETE' }, true);
       // Update the app menu
@@ -380,6 +395,8 @@ const DataManagerProvider = ({ allIcons, children }) => {
       schemaType,
     });
   };
+
+  console.log({ contentTypes, initialData, modifiedData });
 
   return (
     <DataManagerContext.Provider
